SOV-3288: Show the current state of voting (#632)
* Show the current state of voting

* Create thin-geckos-fail.md

* Fix the copy
