+++ .changeset/thin-geckos-fail.md
@@ -1,0 +1,5 @@
+---
+"frontend": patch
+---
+
+SOV-3288: Show the current state of voting
