+++ apps/frontend/src/app/5_pages/BitocracyPage/hooks/useProposalStatus.ts
@@ -1,12 +1,10 @@
 import { useMemo } from 'react';
 
-import { Decimal } from '@sovryn/utils';
-
 import { MS } from '../../../../constants/general';
 import { useBlockNumber } from '../../../../hooks/useBlockNumber';
 import { Proposal } from '../../../../utils/graphql/rsk/generated';
-import { decimalic } from '../../../../utils/math';
 import { ProposalState } from '../BitocracyPage.types';
+import { shouldProposalBeDefeated } from '../BitocracyPage.utils';
 import { GRACE_PERIOD_IN_SECONDS } from '../components/Proposals/Proposals.constants';
 
 export const useProposalStatus = (proposal?: Proposal) => {
@@ -20,24 +18,13 @@ export const useProposalStatus = (proposal?: Proposal) => {
 
     let status: ProposalState;
 
-    const totalVotes = Decimal.fromBigNumberString(proposal.votesFor).add(
-      Decimal.fromBigNumberString(proposal.votesAgainst),
-    );
-
-    const supportPercentage = Decimal.fromBigNumberString(proposal.votesFor)
-      .div(totalVotes)
-      .mul(100);
-
     if (proposal.canceled) {
       status = ProposalState.Canceled;
     } else if (blockNumber <= proposal.endBlock) {
       status = ProposalState.Active;
     } else if (blockNumber <= proposal.startBlock) {
       status = ProposalState.Pending;
-    } else if (
-      supportPercentage.lte(proposal.emittedBy.majorityPercentageVotes) ||
-      decimalic(totalVotes).lt(Decimal.fromBigNumberString(proposal.quorum))
-    ) {
+    } else if (shouldProposalBeDefeated(proposal)) {
       status = ProposalState.Defeated;
     } else if (proposal.eta === 0) {
       status = ProposalState.Succeeded;
